--- test/main.js
@@ -1,488 +1,0 @@
-'use strict'
-
-var test = require("tape")
-
-var h = require("../h.js")
-var diff = require("../diff.js")
-var patch = require("../patch.js")
-var render = require("../create-element.js")
-var Node = require("../vnode/vnode")
-var TextNode = require("../vnode/vtext")
-var version = require("../vnode/version")
-var assertEqualDom = require("./lib/assert-equal-dom.js")
-
-
-
-// VirtualNode tests
-test("Node is a function", function (assert) {
-    assert.equal(typeof Node, "function")
-    assert.end()
-})
-
-test("Node type and version are set", function (assert) {
-    assert.equal(Node.prototype.type, "VirtualNode")
-    assert.deepEqual(Node.prototype.version, version)
-    assert.end()
-})
-
-test("TextNode is a function", function (assert) {
-    assert.equal(typeof TextNode, "function")
-    assert.end()
-})
-
-test("TextNode type and version are set", function (assert) {
-    assert.equal(TextNode.prototype.type, "VirtualText")
-    assert.deepEqual(TextNode.prototype.version, version)
-    assert.end()
-})
-
-// h tests
-
-test("h is a function", function (assert) {
-    assert.equal(typeof h, "function")
-    assert.end()
-})
-
-test("defaults to div node", function (assert) {
-    var node = h()
-    assertNode(assert, node, "DIV")
-    assert.end()
-})
-
-test("can use class selector", function (assert) {
-    var node = h("div.pretty")
-    assertNode(assert, node, "DIV", { className: "pretty" })
-    assert.end()
-})
-
-test("can use non-ascii class selector", function (assert) {
-    var node = h("div.ΑΒΓΔΕΖ")
-    assertNode(assert, node, "DIV", { className: "ΑΒΓΔΕΖ" })
-    assert.end()
-})
-
-test("class selectors combine with className property", function (assert) {
-    var node = h("div.very", { className: "pretty" })
-    assertNode(assert, node, "DIV", { className: "very pretty" })
-    assert.end()
-})
-
-test("can use id selector", function (assert) {
-    var node = h("div#important")
-    assertNode(assert, node, "DIV", { id: "important" })
-    assert.end()
-})
-
-test("can use non-ascii id selector", function (assert) {
-    var node = h("div#ΑΒΓΔΕΖ")
-    assertNode(assert, node, "DIV", { id: "ΑΒΓΔΕΖ" })
-    assert.end()
-})
-
-test("properties id overrides selector id", function (assert) {
-    var node = h("div#very", { id: "important" })
-    assertNode(assert, node, "DIV", { id: "important" })
-    assert.end()
-})
-
-test("defaults to div when using selectors", function (assert) {
-    var node1 = h("#important")
-    var node2 = h(".pretty")
-    var node3 = h("#important.pretty")
-    var node4 = h(".pretty#important")
-
-    assertNode(assert, node1, "DIV", { id: "important" })
-    assertNode(assert, node2, "DIV", { className: "pretty" })
-    assertNode(assert, node3, "DIV", { id: "important", className: "pretty" })
-    assertNode(assert, node4, "DIV", { id: "important", className: "pretty" })
-    assert.end()
-})
-
-test("second argument can be children", function (assert) {
-    var node1 = h("#important.pretty", "test")
-    var node2 = h("#important.pretty", ["test"])
-    var node3 = h("#important.pretty", h("p", "testing"))
-    var node4 = h("#important.pretty", [h("p", "testing")])
-
-    var props = { id: "important", className: "pretty" }
-
-    assertNode(assert, node1, "DIV", props, ["test"])
-    assertNode(assert, node2, "DIV", props, ["test"])
-    assertNode(assert, node3, "DIV", props, [["P", {}, ["testing"]]])
-    assertNode(assert, node4, "DIV", props, [["P", {}, ["testing"]]])
-    assert.end()
-})
-
-test("third argument can be child or children", function (assert) {
-    var node1 = h("#important.pretty", { a: "b" }, "test")
-    var node2 = h("#important.pretty", { a: "b" }, ["test"])
-    var node3 = h("#important.pretty", { a: "b" }, h("p", "testing"))
-    var node4 = h("#important.pretty", { a: "b" }, [h("p", "testing")])
-
-    var props = { a: "b", id: "important", className: "pretty" }
-
-    assertNode(assert, node1, "DIV", props, ["test"])
-    assertNode(assert, node2, "DIV", props, ["test"])
-    assertNode(assert, node3, "DIV", props, [["P", {}, ["testing"]]])
-    assertNode(assert, node4, "DIV", props, [["P", {}, ["testing"]]])
-    assert.end()
-})
-
-function assertNode(assert, node, tagName, properties, children) {
-    properties = properties || {}
-    children = children || []
-
-    assert.ok(node instanceof Node, "node is a VirtualNode")
-    assert.equal(node.tagName, tagName, "tag names are equal")
-    assert.deepEqual(node.properties, properties, "propeties are equal")
-    assert.equal(node.children.length, children.length, "child count equal")
-    for (var i = 0; i < children.length; i++) {
-        var child = children[i]
-
-        if (typeof child === "string") {
-            assert.equal(node.children[i].text, child)
-        } else {
-            assertNode(assert,
-                node.children[i],
-                child[0],
-                child[1],
-                child[2])
-        }
-    }
-}
-
-
-
-// render tests
-test("render is a function", function (assert) {
-    assert.equal(typeof h, "function")
-    assert.end()
-})
-
-test("render text node", function (assert) {
-    var vdom = h("span", "hello")
-    var dom = render(vdom)
-    assert.equal(dom.tagName, "SPAN")
-    assert.notOk(dom.id)
-    assert.notOk(dom.className)
-    assert.equal(dom.childNodes.length, 1)
-    assert.equal(dom.childNodes[0].data, "hello")
-    assert.end()
-})
-
-test("render div", function (assert) {
-    var vdom = h()
-    var dom = render(vdom)
-    assert.notOk(dom.id)
-    assert.notOk(dom.className)
-    assert.equal(dom.tagName, "DIV")
-    assert.equal(dom.childNodes.length, 0)
-    assert.end()
-})
-
-test("node id is applied correctly", function (assert) {
-    var vdom = h("#important")
-    var dom = render(vdom)
-    assert.equal(dom.id, "important")
-    assert.notOk(dom.className)
-    assert.equal(dom.tagName, "DIV")
-    assert.equal(dom.childNodes.length, 0)
-    assert.end()
-})
-
-test("node class name is applied correctly", function (assert) {
-    var vdom = h(".pretty")
-    var dom = render(vdom)
-    assert.notOk(dom.id)
-    assert.equal(dom.className, "pretty")
-    assert.equal(dom.tagName, "DIV")
-    assert.equal(dom.childNodes.length, 0)
-    assert.end()
-})
-
-test("mixture of node/classname applied correctly", function (assert) {
-    var vdom = h("#override.very", { id: "important", className: "pretty"})
-    var dom = render(vdom)
-    assert.equal(dom.id, "important")
-    assert.equal(dom.className, "very pretty")
-    assert.equal(dom.tagName, "DIV")
-    assert.equal(dom.childNodes.length, 0)
-    assert.end()
-})
-
-test("style object is applied correctly", function (assert) {
-    var vdom = h("#important.pretty", { style: {
-        border: "1px solid rgb(0, 0, 0)",
-        padding: "2px"
-    } })
-    var dom = render(vdom)
-    assert.equal(dom.id, "important")
-    assert.equal(dom.className, "pretty")
-    assert.equal(dom.tagName, "DIV")
-    assert.equal(dom.style.border, style("border", "1px solid rgb(0, 0, 0)"))
-    assert.equal(dom.style.padding, style("padding", "2px"))
-    assert.equal(dom.childNodes.length, 0)
-    assert.end()
-})
-
-test("children are added", function (assert) {
-    var vdom = h("div", [
-        h("div", [
-            "just testing",
-            "multiple",
-            h("b", "nodes")
-        ]),
-        "hello",
-        h("span", "test")
-    ])
-
-    var dom = render(vdom)
-
-    assert.equal(dom.childNodes.length, 3)
-
-    var nodes = dom.childNodes
-    assert.equal(nodes.length, 3)
-    assert.equal(nodes[0].tagName, "DIV")
-    assert.equal(nodes[1].data, "hello")
-    assert.equal(nodes[2].tagName, "SPAN")
-
-    var subNodes0 = nodes[0].childNodes
-    assert.equal(subNodes0.length, 3)
-    assert.equal(subNodes0[0].data, "just testing")
-    assert.equal(subNodes0[1].data, "multiple")
-    assert.equal(subNodes0[2].tagName, "B")
-
-    var subNodes0_2 = subNodes0[2].childNodes
-    assert.equal(subNodes0_2.length, 1)
-    assert.equal(subNodes0_2[0].data, "nodes")
-
-    var subNodes2 = nodes[2].childNodes
-    assert.equal(subNodes2.length, 1)
-    assert.equal(subNodes2[0].data, "test")
-    assert.end()
-})
-
-test("incompatible children are ignored", function (assert) {
-    var vdom = h("#important.pretty", {
-        style: {
-            "cssText": "color: red;"
-        }
-    }, [
-        null
-    ])
-    var dom = render(vdom)
-    assert.equal(dom.id, "important")
-    assert.equal(dom.className, "pretty")
-    assert.equal(dom.tagName, "DIV")
-    assert.equal(dom.style.cssText, style("cssText", "color: red;"))
-    assert.equal(dom.childNodes.length, 0)
-    assert.end()
-})
-
-test("injected document object is used", function (assert) {
-    var vdom = h("div", "hello")
-    var count = 0
-    var doc = {
-        createElement: function createElement(tagName) {
-            assert.equal(tagName, "DIV")
-            count++
-            return { tagName: "DIV", appendChild: function (t) {
-                assert.equal(t, "hello")
-                count++
-            } }
-        },
-        createTextNode: function createTextNode(text) {
-            assert.equal(text, "hello")
-            count++
-            return text
-        }
-    }
-    render(vdom, { document: doc })
-    assert.equal(count, 3)
-    assert.end()
-})
-
-test("injected warning is used", function (assert) {
-    var badObject = {}
-    var vdom = h("#important.pretty", {
-        style: {
-            cssText: "color: red;"
-        }
-    })
-
-    vdom.children = [
-        badObject, null
-    ]
-
-    var i = 0
-    function warn(warning, node) {
-        assert.equal(warning, "Item is not a valid virtual dom node")
-
-        if (i === 0) {
-            assert.equal(node, badObject)
-        } else if (i === 1) {
-            assert.equal(node, null)
-        } else {
-            assert.error("Too many warnings")
-        }
-
-        i++
-    }
-
-    var dom = render(vdom, { warn: warn })
-    assert.equal(dom.id, "important")
-    assert.equal(dom.className, "pretty")
-    assert.equal(dom.tagName, "DIV")
-    assert.equal(dom.style.cssText, style("cssText", "color: red;"))
-    assert.equal(dom.childNodes.length, 0)
-    assert.equal(i, 2)
-    assert.end()
-})
-
-// Complete patch tests
-test("textnode update test", function (assert) {
-    var hello = h("div", "hello")
-    var goodbye = h("div", "goodbye")
-    var rootNode = render(hello)
-    var equalNode = render(goodbye)
-    var patches = diff(hello, goodbye)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-test("textnode replace test", function (assert) {
-    var hello = h("div", "hello")
-    var goodbye = h("div", [h("span", "goodbye")])
-    var rootNode = render(hello)
-    var equalNode = render(goodbye)
-    var patches = diff(hello, goodbye)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-test("textnode insert test", function (assert) {
-    var hello = h("div", "hello")
-    var again = h("span", ["hello", "again"])
-    var rootNode = render(hello)
-    var equalNode = render(again)
-    var patches = diff(hello, again)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-test("textnode remove", function (assert) {
-    var again = h("span", ["hello", "again"])
-    var hello = h("div", "hello")
-    var rootNode = render(again)
-    var equalNode = render(hello)
-    var patches = diff(again, hello)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-test("dom node update test", function (assert) {
-    var hello = h("div.hello", "hello")
-    var goodbye = h("div.goodbye", "goodbye")
-    var rootNode = render(hello)
-    var equalNode = render(goodbye)
-    var patches = diff(hello, goodbye)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-test("dom node replace test", function (assert) {
-    var hello = h("div", "hello")
-    var goodbye = h("span", "goodbye")
-    var rootNode = render(hello)
-    var equalNode = render(goodbye)
-    var patches = diff(hello, goodbye)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-test("dom node insert", function (assert) {
-    var hello = h("div", [h("span", "hello")])
-    var again = h("div", [h("span", "hello"), h("span", "again")])
-    var rootNode = render(hello)
-    var equalNode = render(again)
-    var patches = diff(hello, again)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-test("dom node remove", function (assert) {
-    var hello = h("div", [h("span", "hello")])
-    var again = h("div", [h("span", "hello"), h("span", "again")])
-    var rootNode = render(again)
-    var equalNode = render(hello)
-    var patches = diff(again, hello)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-    assert.end()
-})
-
-
-test("reuse dom node without breaking", function (assert) {
-    var hSpan = h("span", "hello")
-    var hello = h("div", [hSpan, hSpan, hSpan])
-    var goodbye = h("div", [h("span", "hello"), hSpan, h("span", "goodbye")])
-    var rootNode = render(hello)
-    var equalNode = render(goodbye)
-    var patches = diff(hello, goodbye)
-    var newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, equalNode)
-
-    // Undo the rendering with new trees
-    hello = h("div", [hSpan, hSpan, hSpan])
-    goodbye = h("div", [h("span", "hello"), hSpan, h("span", "goodbye")])
-    patches = diff(goodbye, hello)
-    newRoot = patch(rootNode, patches)
-    assertEqualDom(assert, newRoot, rootNode)
-
-    assert.end()
-})
-
-test("Allow empty textnode", function (assert) {
-    var empty = h("span", "")
-    var rootNode = render(empty)
-    assert.equal(rootNode.childNodes.length, 1)
-    assert.equal(rootNode.childNodes[0].data, "")
-    assert.end()
-})
-
-test("Can replace vnode with vtext", function (assert) {
-
-    var leftNode = h("div", h("div"))
-    var rightNode = h("div", "text")
-
-    var rootNode = render(leftNode)
-
-    assert.equal(rootNode.childNodes.length, 1)
-    assert.equal(rootNode.childNodes[0].nodeType, 1)
-
-    var patches = diff(leftNode, rightNode)
-
-    var newRoot = patch(rootNode, patches)
-
-    assert.equal(newRoot, rootNode)
-
-    assert.equal(newRoot.childNodes.length, 1)
-    assert.equal(newRoot.childNodes[0].nodeType, 3)
-
-    assert.end()
-})
-
-
-// Safely translates style values using the DOM in the browser
-function style(name, value) {
-    var node = render(h())
-    node.style[name] = value
-    return node.style[name]
-}
